@@ -1,9 +1,19 @@
 package jiguang.chat.activity;
 
+import android.Manifest;
 import android.os.Bundle;
 import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentManager;
 import android.view.KeyEvent;
+import android.widget.Toast;
+
+import com.yanzhenjie.alertdialog.AlertDialog;
+import com.yanzhenjie.permission.AndPermission;
+import com.yanzhenjie.permission.PermissionListener;
+import com.yanzhenjie.permission.RationaleListener;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import cn.jiguang.api.JCoreInterface;
 import jiguang.chat.R;
@@ -13,19 +23,49 @@ import jiguang.chat.view.MainView;
 public class MainActivity extends FragmentActivity {
     private MainController mMainController;
     private MainView mMainView;
+    private List<String> permissions;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-
+        requestPermission();
         mMainView = (MainView) findViewById(R.id.main_view);
         mMainView.initModule();
         mMainController = new MainController(mMainView, this);
 
         mMainView.setOnClickListener(mMainController);
         mMainView.setOnPageChangeListener(mMainController);
+    }
+
+    private void requestPermission() {
+        permissions = new ArrayList<>();
+        permissions.add(Manifest.permission.READ_EXTERNAL_STORAGE);
+        permissions.add(Manifest.permission.WRITE_EXTERNAL_STORAGE);
+        permissions.add(Manifest.permission.CAMERA);
+        permissions.add(Manifest.permission.ACCESS_COARSE_LOCATION);
+        permissions.add(Manifest.permission.ACCESS_COARSE_LOCATION);
+        permissions.add(Manifest.permission.ACCESS_FINE_LOCATION);
+        permissions.add(Manifest.permission.ACCESS_WIFI_STATE);
+        permissions.add(Manifest.permission.RECORD_AUDIO);
+        if (AndPermission.hasPermission(this, permissions)) {
+            return;
+        }else {
+            AndPermission.with(this)
+                    .requestCode(100)
+                    .permission(Manifest.permission.READ_EXTERNAL_STORAGE,
+                            Manifest.permission.WRITE_EXTERNAL_STORAGE,
+                            Manifest.permission.CAMERA,
+                            Manifest.permission.ACCESS_COARSE_LOCATION,
+                            Manifest.permission.ACCESS_COARSE_LOCATION,
+                            Manifest.permission.ACCESS_FINE_LOCATION,
+                            Manifest.permission.ACCESS_WIFI_STATE,
+                            Manifest.permission.RECORD_AUDIO)
+                    .rationale(rationaleListener)
+                    .callback(listener)
+                    .start();
+        }
     }
 
     public FragmentManager getSupportFragmentManger() {
@@ -53,4 +93,33 @@ public class MainActivity extends FragmentActivity {
     public boolean onKeyDown(int keyCode, KeyEvent event) {
         return keyCode == KeyEvent.KEYCODE_BACK;
     }
+
+
+    private PermissionListener listener = new PermissionListener() {
+        @Override
+        public void onSucceed(int requestCode, List<String> grantedPermissions) {
+            // Successfully.
+            if(requestCode == 100) {
+                Toast.makeText(MainActivity.this, "成功", Toast.LENGTH_SHORT).show();
+            }
+        }
+
+        @Override
+        public void onFailed(int requestCode, List<String> deniedPermissions) {
+            // Failure.
+
+        }
+    };
+
+    private RationaleListener rationaleListener = (requestCode, rationale) -> {
+        AlertDialog.newBuilder(this)
+                .setMessage("正常使用JChat需要您打开相关权限.")
+                .setPositiveButton("是", (dialog, which) -> {
+                    rationale.resume();
+                })
+                .setNegativeButton("否", (dialog, which) -> {
+                    rationale.cancel();
+                }).show();
+    };
+
 }
