@@ -17,19 +17,25 @@ import cn.jpush.im.android.api.callback.GetGroupInfoCallback;
 import cn.jpush.im.android.api.callback.GetUserInfoCallback;
 import cn.jpush.im.android.api.callback.GetUserInfoListCallback;
 import cn.jpush.im.android.api.event.ContactNotifyEvent;
+import cn.jpush.im.android.api.event.GroupApprovalEvent;
+import cn.jpush.im.android.api.event.GroupApprovalRefuseEvent;
 import cn.jpush.im.android.api.model.Conversation;
 import cn.jpush.im.android.api.model.GroupInfo;
 import cn.jpush.im.android.api.model.UserInfo;
 import cn.jpush.im.android.eventbus.EventBus;
+import cn.jpush.im.android.utils.JsonUtil;
 import jiguang.chat.R;
 import jiguang.chat.application.JGApplication;
 import jiguang.chat.controller.ContactsController;
 import jiguang.chat.database.FriendEntry;
 import jiguang.chat.database.FriendRecommendEntry;
+import jiguang.chat.database.GroupApplyEntry;
+import jiguang.chat.database.RefuseGroupEntry;
 import jiguang.chat.database.UserEntry;
 import jiguang.chat.entity.Event;
 import jiguang.chat.entity.EventType;
 import jiguang.chat.entity.FriendInvitation;
+import jiguang.chat.entity.GroupApplyInvitation;
 import jiguang.chat.utils.SharePreferenceManager;
 import jiguang.chat.utils.ThreadUtil;
 import jiguang.chat.utils.pinyin.HanziToPinyin;
@@ -44,7 +50,6 @@ public class ContactsFragment extends BaseFragment {
     private ContactsView mContactsView;
     private ContactsController mContactsController;
     private Activity mContext;
-
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -237,6 +242,92 @@ public class ContactsFragment extends BaseFragment {
                 }
             }
         }
+    }
+
+    //群主收到群组验证事件
+    public void onEvent(GroupApprovalEvent event) {
+        final UserEntry user = JGApplication.getUserEntry();
+        GroupApprovalEvent.Type type = event.getType();
+        long gid = event.getGid();
+        event.getFromUserInfo(new GetUserInfoCallback() {
+            @Override
+            public void gotResult(int i, String s, UserInfo fromUserInfo) {
+                if (i == 0) {
+                    event.getApprovalUserInfoList(new GetUserInfoListCallback() {
+                        @Override
+                        public void gotResult(int i, String s, List<UserInfo> list) {
+                            if (i == 0) {
+                                GroupApplyEntry entry;
+                                //邀请,from是邀请方
+                                if (type.equals(GroupApprovalEvent.Type.invited_into_group)) {
+                                    entry = GroupApplyEntry.getEntry(user, list.get(0).getUserName(), list.get(0).getAppKey());
+                                    if (entry != null) {
+                                        entry.delete();
+                                    }
+                                    if (fromUserInfo.getAvatar() != null) {
+                                        entry = new GroupApplyEntry(fromUserInfo.getUserName(), list.get(0).getUserName(), fromUserInfo.getAppKey(),
+                                                list.get(0).getAvatarFile().getPath(), fromUserInfo.getDisplayName(), list.get(0).getDisplayName(),
+                                                null, GroupApplyInvitation.INVITED.getValue(), JsonUtil.toJson(event), gid + "",
+                                                user, 0, 0);//邀请type=0
+                                    } else {
+                                        entry = new GroupApplyEntry(fromUserInfo.getUserName(), list.get(0).getUserName(), fromUserInfo.getAppKey(),
+                                                null, fromUserInfo.getDisplayName(), list.get(0).getDisplayName(),
+                                                null, GroupApplyInvitation.INVITED.getValue(), JsonUtil.toJson(event), gid + "",
+                                                user, 0, 0);//邀请type=0
+                                    }
+                                } else {
+                                    entry = GroupApplyEntry.getEntry(user, fromUserInfo.getUserName(), fromUserInfo.getAppKey());
+                                    if (entry != null) {
+                                        entry.delete();
+                                    }
+                                    if (fromUserInfo.getAvatar() != null) {
+                                        entry = new GroupApplyEntry(list.get(0).getUserName(), list.get(0).getUserName(), list.get(0).getAppKey(),
+                                                list.get(0).getAvatarFile().getPath(), list.get(0).getDisplayName(), list.get(0).getDisplayName(),
+                                                event.getReason(), GroupApplyInvitation.INVITED.getValue(), JsonUtil.toJson(event), gid + "",
+                                                user, 0, 1);//申请type=1
+                                    } else {
+                                        entry = new GroupApplyEntry(list.get(0).getUserName(), list.get(0).getUserName(), list.get(0).getAppKey(),
+                                                null, fromUserInfo.getDisplayName(), list.get(0).getDisplayName(),
+                                                event.getReason(), GroupApplyInvitation.INVITED.getValue(), JsonUtil.toJson(event), gid + "",
+                                                user, 0, 1);//申请type=1
+                                    }
+                                }
+                                entry.save();
+                            }
+                        }
+                    });
+
+                }
+            }
+        });
+    }
+
+    //收到被拒绝事件
+    public void onEvent(GroupApprovalRefuseEvent event) {
+        final UserEntry user = JGApplication.getUserEntry();
+        long gid = event.getGid();
+        event.getToUserInfoList(new GetUserInfoListCallback() {
+            @Override
+            public void gotResult(int i, String s, List<UserInfo> list) {
+                if (i == 0) {
+                    String userName = list.get(0).getUserName();
+                    String displayName = list.get(0).getDisplayName();
+                    String appKey = list.get(0).getAppKey();
+                    String path = null;
+                    if (list.get(0).getAvatar() != null) {
+                        path = list.get(0).getAvatarFile().getPath();
+                    }
+                    RefuseGroupEntry groupEntry = RefuseGroupEntry.getEntry(user, userName, appKey);
+                    if (groupEntry != null) {
+                        groupEntry.delete();
+                    }
+                    groupEntry = new RefuseGroupEntry(user, userName, displayName, gid + "", appKey, path);
+                    groupEntry.save();
+
+                }
+            }
+        });
+
     }
 
     private String getLetter(String name) {
