@@ -7,7 +7,6 @@ import cn.jpush.im.android.api.model.UserInfo;
 import jiguang.chat.R;
 import jiguang.chat.activity.FriendInfoActivity;
 import jiguang.chat.activity.FriendSettingActivity;
-import jiguang.chat.view.FriendInfoView;
 
 /**
  * Created by ${chenyn} on 2017/3/24.
@@ -16,9 +15,11 @@ import jiguang.chat.view.FriendInfoView;
 public class FriendInfoController implements View.OnClickListener {
     private FriendInfoActivity mContext;
     private UserInfo friendInfo;
+    private int flags;
 
-    public FriendInfoController(FriendInfoView friendInfoView, FriendInfoActivity context) {
+    public FriendInfoController(int flags, FriendInfoActivity context) {
         this.mContext = context;
+        this.flags = flags;
     }
 
     @Override
@@ -31,7 +32,8 @@ public class FriendInfoController implements View.OnClickListener {
                 mContext.startBrowserAvatar();
                 break;
             case R.id.jmui_commit_btn:
-                Intent intent = new Intent(mContext, FriendSettingActivity.class);
+                Intent intent = new Intent();
+                intent.setClass(mContext, FriendSettingActivity.class);
                 intent.putExtra("userName", friendInfo.getUserName());
                 intent.putExtra("noteName", friendInfo.getNotename());
                 mContext.startActivity(intent);
