@@ -26,10 +26,10 @@ public class MainView extends RelativeLayout {
 
     public void initModule() {
         mBtnListID = new int[] {
-                R.id.actionbar_msg_btn, R.id.actionbar_contact_btn, R.id.actionbar_me_btn
-        };
-        mBtnList = new Button[3];
-        for (int i = 0; i < 3; i++) {
+                R.id.actionbar_msg_btn, R.id.actionbar_contact_btn,
+                R.id.actionbar_chatroom_btn, R.id.actionbar_me_btn};
+        mBtnList = new Button[mBtnListID.length];
+        for (int i = 0; i < mBtnListID.length; i++) {
             mBtnList[i] = (Button) findViewById(mBtnListID[i]);
         }
         mViewContainer = (ScrollControlViewPager) findViewById(R.id.viewpager);
@@ -56,7 +56,7 @@ public class MainView extends RelativeLayout {
     }
 
     public void setButtonColor(int index) {
-        for (int i = 0; i < 3; i++) {
+        for (int i = 0; i < mBtnListID.length; i++) {
             if (index == i) {
                 mBtnList[i].setSelected(true);
                 mBtnList[i].setTextColor(getResources().getColor(R.color.actionbar_pres_color));
