@@ -1,7 +1,7 @@
 package jiguang.chat.view;
 
 import android.view.View;
-import android.widget.LinearLayout;
+import android.widget.RelativeLayout;
 
 import jiguang.chat.R;
 
@@ -9,20 +9,22 @@ import jiguang.chat.R;
 public class MenuItemView {
 
     private View mView;
-    private LinearLayout mCreateGroupLl;
-    private LinearLayout mAddFriendLl;
-    private LinearLayout mSendMsgLl;
-    private LinearLayout mLl_saoYiSao;
+    private RelativeLayout mCreateGroupLl;
+    private RelativeLayout mAddFriendLl;
+    private RelativeLayout mSendMsgLl;
+    private RelativeLayout mLl_saoYiSao;
+    private RelativeLayout mAdd_open_group;
 
     public MenuItemView(View view) {
         this.mView = view;
     }
 
     public void initModule() {
-        mCreateGroupLl = (LinearLayout) mView.findViewById(R.id.create_group_ll);
-        mAddFriendLl = (LinearLayout) mView.findViewById(R.id.add_friend_with_confirm_ll);
-        mSendMsgLl = (LinearLayout) mView.findViewById(R.id.send_message_ll);
-        mLl_saoYiSao = (LinearLayout) mView.findViewById(R.id.ll_saoYiSao);
+        mCreateGroupLl = mView.findViewById(R.id.create_group_ll);
+        mAddFriendLl = mView.findViewById(R.id.add_friend_with_confirm_ll);
+        mSendMsgLl = mView.findViewById(R.id.send_message_ll);
+        mLl_saoYiSao = mView.findViewById(R.id.ll_saoYiSao);
+        mAdd_open_group = mView.findViewById(R.id.add_open_group);
     }
 
     public void setListeners(View.OnClickListener listener) {
@@ -30,6 +32,7 @@ public class MenuItemView {
         mAddFriendLl.setOnClickListener(listener);
         mSendMsgLl.setOnClickListener(listener);
         mLl_saoYiSao.setOnClickListener(listener);
+        mAdd_open_group.setOnClickListener(listener);
     }
 
     public void showAddFriendDirect() {
